server/bot/requirements.txt
flask
openai
python-dotenv
pymongo
flask-cors
bcrypt
PyJWT
pypdf
langchain
chromadb
sentence-transformers
tiktoken
langchain-huggingface